Initial Commit
1. Splashscene: I need the background only, not with the text.
2. Create the bmfont using http://kvazars.com/littera/ , I tried to find
the font you used but couldn't. Note that you can edit the png file
after created, but don't change the position of letters inside. I
attached a sample bm font set below
3. Player info: I edited the information input board, but as we
discussed, I need the background in landscape (1536x2048). Currently I'm
using the background of loading screen, which is landscape and same size
as temporary.  There's no button also
4. There's no idle animation for characters.
5. There's no assets for main menu (where we select scenes to play)
6. There's no icon for the game
7. The portrait is REALLY BIG.

I've made the joystick, but for 4-dimensions. Creating new one with 2
dimensions for boss-fighting.
Program flows:

Splash Scene: check if User is created. If not, move to PlayerInfo Scene
to create, or else move directly to the main menu.
PlayerInfo: currently validate if the name and age is empty or not, and
the age is numeric only.
Main Menu: Choose the game to go to, currently only the Fighting.
Load Scene: get the resources to load, then auto redirect to the
dedicated scene.